do not upload pkls if environment variable UPLOAD_PKL!=1

+++ darts-models/upload_pkl.sh
@@ -2,9 +2,17 @@
 commit=$4
 odls=$5
 py=$6
+
+if [ "$UPLOAD_PKL" != 1 ]; then
+	exit
+fi
+
 fname="$commit"_"$odls"_"$py.tar.gz"
 echo $fname
+
 tar -czf $fname ./*/*.pkl
+
 #smbclient -U $2%$3 //$1/opendarts-private-artifacts -c "prompt OFF;mkdir pkl"
 smbclient -U $2%$3 //$1/opendarts-private-artifacts -c "put $fname" -D=pkl
+
 rm $fname
